implemented some more functionalities (booking button on a bookPage and Calendar interactions)

+++ php/utilis/update_profile.php
@@ -2,7 +2,6 @@
 require_once '../db/Bootstrap.php';
 
 
-// Check if the request is coming through
 if(isset($_POST["nome"])){ 
     
     $email = $_SESSION["email"];
@@ -10,12 +9,8 @@ if(isset($_POST["nome"])){
     $cognome = $_POST["cognome"];
     $corso = $_POST["corso"];
     $anno = $_POST["anno"];
-    
-    // Default: Keep existing image if no new one is uploaded
-    // (You might want to fetch the current image from DB first)
     $finalImagePath = $_POST['current_image_hidden'] ?? $_SESSION["img"]; 
-
-    // Handle the File Upload
+    
     if (isset($_FILES['profile_pic']) && $_FILES['profile_pic']['error'] === UPLOAD_ERR_OK) {
         $uploadDir = 'uploads/';
         $fileName = time() . "_" . basename($_FILES['profile_pic']['name']);
@@ -24,7 +19,9 @@ if(isset($_POST["nome"])){
         if (move_uploaded_file($_FILES['profile_pic']['tmp_name'], '../'.$targetPath)) {
             $finalImagePath = $targetPath;
         }
-        unlink('../'.$_POST['current_image_hidden']);
+        if($_POST['current_image_hidden']!='default_avatar.png'){
+            unlink('../'.$_POST['current_image_hidden']);
+        }
     }
 
     if($dbh->updateUserInfos($email, $nome, $cognome, $corso, $anno, $finalImagePath)["success"]){
